Fixed #26840 -- Added test.utils.setup/teardown_databases().
